// gql/graphql.ts
export type Maybe<T> = T | null;
// ...
export type Asset = Entity & Node & {
  __typename?: 'Asset';
  /** The time the document was created */
  createdAt: Scalars['DateTime']['output'];
  /** User that created this document */
  createdBy?: Maybe<User>;
  /** Get the document in other stages */
  documentInStages: Array<Asset>;
  /** The file name */
  fileName: Scalars['String']['output'];
  /** The file handle */
  handle: Scalars['String']['output'];
  /** The height of the file */
  height?: Maybe<Scalars['Float']['output']>;
  /** List of Asset versions */
  history: Array<Version>;
  /** The unique identifier */
  id: Scalars['ID']['output'];
  imageProductFocus: Array<ProductFocus>;
  /** System Locale field */
  locale: Locale;
  /** Get the other localizations for this document */
  localizations: Array<Asset>;
  /** The mime type of the file */
  mimeType?: Maybe<Scalars['String']['output']>;
  ogImagePage: Array<Page>;
  ogImagePdp: Array<Pdp>;
  /** The time the document was published. Null on documents in draft stage. */
  publishedAt?: Maybe<Scalars['DateTime']['output']>;
  /** User that last published this document */
  publishedBy?: Maybe<User>;
  scheduledIn: Array<ScheduledOperation>;
  /** The file size */
  size?: Maybe<Scalars['Float']['output']>;
  /** System stage field */
  stage: Stage;
  /** The time the document was updated */
  updatedAt: Scalars['DateTime']['output'];
  /** User that last updated this document */
  updatedBy?: Maybe<User>;
  /** Get the url for the asset with provided transformations applied. */
  url: Scalars['String']['output'];
  /** The file width */
  width?: Maybe<Scalars['Float']['output']>;
};
// ...
export type AssetConnection = {
  __typename?: 'AssetConnection';
  aggregate: Aggregate;
  /** A list of edges. */
  edges: Array<AssetEdge>;
  /** Information to aid in pagination. */
  pageInfo: PageInfo;
};
// ...
export type DocumentVersion = {
  __typename?: 'DocumentVersion';
  createdAt: Scalars['DateTime']['output'];
  data?: Maybe<Scalars['Json']['output']>;
  id: Scalars['ID']['output'];
  revision: Scalars['Int']['output'];
  stage: Stage;
};
// ...
export type Entity = {
  /** The id of the object. */
  id: Scalars['ID']['output'];
  /** The Stage of an object */
  stage: Stage;
};
// ...
export type FederateThisSkincre_Image = {
  __typename?: 'FederateThisSkincre_Image';
  alt?: Maybe<Scalars['String']['output']>;
  url: Scalars['String']['output'];
};

export type FederateThisSkincre_Product = {
  __typename?: 'FederateThisSkincre_Product';
  description?: Maybe<Scalars['String']['output']>;
  id: Scalars['Int']['output'];
  images?: Maybe<Array<Maybe<FederateThisSkincre_Image>>>;
  ingredients?: Maybe<Scalars['String']['output']>;
  name: Scalars['String']['output'];
  price?: Maybe<Scalars['Int']['output']>;
  shortDescription?: Maybe<Scalars['String']['output']>;
  slug: Scalars['String']['output'];
  stock?: Maybe<Scalars['Int']['output']>;
};
// ...
export enum Locale {
  /** System locale */
  En = 'en'
}
// ...
export type Node = {
  /** The id of the object. */
  id: Scalars['ID']['output'];
  /** The Stage of an object */
  stage: Stage;
};
// ...
export type PageConnection = {
  __typename?: 'PageConnection';
  aggregate: Aggregate;
  /** A list of edges. */
  edges: Array<PageEdge>;
  /** Information to aid in pagination. */
  pageInfo: PageInfo;
};
// ...
export type Pdp = Entity & Node & {
  __typename?: 'Pdp';
  components: Array<PdpcomponentsUnion>;
  /** The time the document was created */
  createdAt: Scalars['DateTime']['output'];
  /** User that created this document */
  createdBy?: Maybe<User>;
  description?: Maybe<Scalars['String']['output']>;
  /** Get the document in other stages */
  documentInStages: Array<Pdp>;
  /** List of Pdp versions */
  history: Array<Version>;
  /** The unique identifier */
  id: Scalars['ID']['output'];
  ogImage?: Maybe<Asset>;
  product?: Maybe<FederateThisSkincre_Product>;
  productId?: Maybe<Scalars['String']['output']>;
  /** The time the document was published. Null on documents in draft stage. */
  publishedAt?: Maybe<Scalars['DateTime']['output']>;
  /** User that last published this document */
  publishedBy?: Maybe<User>;
  scheduledIn: Array<ScheduledOperation>;
  slug?: Maybe<Scalars['String']['output']>;
  /** System stage field */
  stage: Stage;
  title?: Maybe<Scalars['String']['output']>;
  /** The time the document was updated */
  updatedAt: Scalars['DateTime']['output'];
  /** User that last updated this document */
  updatedBy?: Maybe<User>;
};
// ...
export type PdpConnection = {
  __typename?: 'PdpConnection';
  aggregate: Aggregate;
  /** A list of edges. */
  edges: Array<PdpEdge>;
  /** Information to aid in pagination. */
  pageInfo: PageInfo;
};
// ...
export type PdpcomponentsUnion = ProductList | Routine | Tutorial;
// ...
export type ProductFocusConnection = {
  __typename?: 'ProductFocusConnection';
  aggregate: Aggregate;
  /** A list of edges. */
  edges: Array<ProductFocusEdge>;
  /** Information to aid in pagination. */
  pageInfo: PageInfo;
};
// ...
export type ProductList = Entity & {
  __typename?: 'ProductList';
  /** The unique identifier */
  id: Scalars['ID']['output'];
  relatedProductList?: Maybe<RelatedProductList>;
  /** System stage field */
  stage: Stage;
  title?: Maybe<Scalars['String']['output']>;
};
// ...
export type Query = {
  __typename?: 'Query';
  /** Retrieve a single asset */
  asset?: Maybe<Asset>;
  /** Retrieve document version */
  assetVersion?: Maybe<DocumentVersion>;
  /** Retrieve multiple assets */
  assets: Array<Asset>;
  /** Retrieve multiple assets using the Relay connection interface */
  assetsConnection: AssetConnection;
  /** Fetches an object given its ID */
  entities?: Maybe<Array<Entity>>;
  /** Fetches an object given its ID */
  node?: Maybe<Node>;
  /** Retrieve a single page */
  page?: Maybe<Page>;
  /** Retrieve document version */
  pageVersion?: Maybe<DocumentVersion>;
  /** Retrieve multiple pages */
  pages: Array<Page>;
  /** Retrieve multiple pages using the Relay connection interface */
  pagesConnection: PageConnection;
  /** Retrieve a single pdp */
  pdp?: Maybe<Pdp>;
  /** Retrieve document version */
  pdpVersion?: Maybe<DocumentVersion>;
  /** Retrieve multiple pdps */
  pdps: Array<Pdp>;
  /** Retrieve multiple pdps using the Relay connection interface */
  pdpsConnection: PdpConnection;
  /** Retrieve multiple productFoci */
  productFoci: Array<ProductFocus>;
  /** Retrieve multiple productFoci using the Relay connection interface */
  productFociConnection: ProductFocusConnection;
  /** Retrieve a single productFocus */
  productFocus?: Maybe<ProductFocus>;
  /** Retrieve document version */
  productFocusVersion?: Maybe<DocumentVersion>;
  products?: Maybe<Array<FederateThisSkincre_Product>>;
  /** Retrieve a single relatedProductList */
  relatedProductList?: Maybe<RelatedProductList>;
  /** Retrieve document version */
  relatedProductListVersion?: Maybe<DocumentVersion>;
  /** Retrieve multiple relatedProductLists */
  relatedProductLists: Array<RelatedProductList>;
  /** Retrieve multiple relatedProductLists using the Relay connection interface */
  relatedProductListsConnection: RelatedProductListConnection;
  /** Retrieve a single scheduledOperation */
  scheduledOperation?: Maybe<ScheduledOperation>;
  /** Retrieve multiple scheduledOperations */
  scheduledOperations: Array<ScheduledOperation>;
  /** Retrieve multiple scheduledOperations using the Relay connection interface */
  scheduledOperationsConnection: ScheduledOperationConnection;
  /** Retrieve a single scheduledRelease */
  scheduledRelease?: Maybe<ScheduledRelease>;
  /** Retrieve multiple scheduledReleases */
  scheduledReleases: Array<ScheduledRelease>;
  /** Retrieve multiple scheduledReleases using the Relay connection interface */
  scheduledReleasesConnection: ScheduledReleaseConnection;
  /** Retrieve a single skncreProduct */
  skncreProduct?: Maybe<SkncreProduct>;
  /** Retrieve document version */
  skncreProductVersion?: Maybe<DocumentVersion>;
  /** Retrieve multiple skncreProducts */
  skncreProducts: Array<SkncreProduct>;
  /** Retrieve multiple skncreProducts using the Relay connection interface */
  skncreProductsConnection: SkncreProductConnection;
  /** Retrieve a single user */
  user?: Maybe<User>;
  /** Retrieve multiple users */
  users: Array<User>;
  /** Retrieve multiple users using the Relay connection interface */
  usersConnection: UserConnection;
};
// ...
export type RelatedProductList = Entity & Node & {
  __typename?: 'RelatedProductList';
  /** The time the document was created */
  createdAt: Scalars['DateTime']['output'];
  /** User that created this document */
  createdBy?: Maybe<User>;
  /** Get the document in other stages */
  documentInStages: Array<RelatedProductList>;
  /** List of RelatedProductList versions */
  history: Array<Version>;
  /** The unique identifier */
  id: Scalars['ID']['output'];
  /** The time the document was published. Null on documents in draft stage. */
  publishedAt?: Maybe<Scalars['DateTime']['output']>;
  /** User that last published this document */
  publishedBy?: Maybe<User>;
  relatedProductId?: Maybe<Scalars['String']['output']>;
  relatedProducts?: Maybe<FederateThisSkincre_Related>;
  scheduledIn: Array<ScheduledOperation>;
  /** System stage field */
  stage: Stage;
  /** The time the document was updated */
  updatedAt: Scalars['DateTime']['output'];
  /** User that last updated this document */
  updatedBy?: Maybe<User>;
};
// ...
export type Routine = Entity & {
  __typename?: 'Routine';
  chapeau?: Maybe<Scalars['String']['output']>;
  cta?: Maybe<Scalars['String']['output']>;
  description?: Maybe<Scalars['String']['output']>;
  /** The unique identifier */
  id: Scalars['ID']['output'];
  image?: Maybe<Asset>;
  /** System stage field */
  stage: Stage;
  title?: Maybe<Scalars['String']['output']>;
  url?: Maybe<Scalars['String']['output']>;
};
// ...
export type Tutorial = Entity & {
  __typename?: 'Tutorial';
  /** The unique identifier */
  id: Scalars['ID']['output'];
  image?: Maybe<Asset>;
  items: Array<TutorialItem>;
  /** System stage field */
  stage: Stage;
  title?: Maybe<Scalars['String']['output']>;
};
// ...
export type TutorialItem = Entity & {
  __typename?: 'TutorialItem';
  /** The unique identifier */
  id: Scalars['ID']['output'];
  /** System stage field */
  stage: Stage;
  text?: Maybe<Scalars['String']['output']>;
};
// ...
export type PdpQuery = { __typename?: 'Query', pdp?: { __typename: 'Pdp', id: string, slug?: string | null, title?: string | null, description?: string | null, ogImage?: { __typename?: 'Asset', url: string } | null, components: Array<{ __typename: 'ProductList', title?: string | null, relatedProductList?: { __typename?: 'RelatedProductList', relatedProductId?: string | null, relatedProducts?: { __typename?: 'FederateThisSkincre_related', products?: Array<{ __typename?: 'FederateThisSkincre_Product', description?: string | null, id: number, ingredients?: string | null, name: string, price?: number | null, shortDescription?: string | null, slug: string, stock?: number | null, images?: Array<{ __typename?: 'FederateThisSkincre_Image', alt?: string | null, url: string } | null> | null }> | null } | null } | null } | { __typename: 'Routine', id: string, chapeau?: string | null, cta?: string | null, description?: string | null, title?: string | null, url?: string | null, image?: { __typename?: 'Asset', url: string } | null } | { __typename: 'Tutorial', id: string, title?: string | null, image?: { __typename?: 'Asset', url: string } | null, items: Array<{ __typename: 'TutorialItem', text?: string | null }> }>, product?: { __typename?: 'FederateThisSkincre_Product', id: number, slug: string, name: string, price?: number | null, ingredients?: string | null, shortDescription?: string | null, description?: string | null, stock?: number | null, images?: Array<{ __typename?: 'FederateThisSkincre_Image', alt?: string | null, url: string } | null> | null } | null } | null };